Feature/category

+++ src/pages/Settings/hooks/useCategorySettings.ts
@@ -2,14 +2,15 @@ import { useState } from "react";
 import { deleteCategory } from "@features/Category/CategoryActions";
 import { Modal, notification } from "antd/lib";
 import { useAppDispatch } from "src/store";
+import { deleteMainCategory, deleteSubCategory } from "@services/categories";
 
 type TdeleteCandidate = { name: string; _id: string } | any;
 
 interface CategorySettingsProps {
-  onDeleteCandidate: (category: TdeleteCandidate) => void;
+  type: string;
 }
 
-const useCategorySettings = ({ onDeleteCandidate }: CategorySettingsProps) => {
+const useCategorySettings = ({ type }: CategorySettingsProps) => {
   const dispatch = useAppDispatch();
   const [deleteConfirmationOpen, setDeleteConfirmationOpen] = useState(false);
   const [deleteCandidate, setDeleteCandidate] = useState<TdeleteCandidate>("");
@@ -22,9 +23,8 @@ const useCategorySettings = ({ onDeleteCandidate }: CategorySettingsProps) => {
 
   const handleDeleteConfirm = async (ref) => {
     try {
-      if (deleteCandidate) {
+      if (type === "category" && deleteCandidate) {
         dispatch(deleteCategory(deleteCandidate._id));
-        onDeleteCandidate(deleteCandidate);
         setDeleteConfirmationOpen(false);
         ref.current?.reload();
 
@@ -33,11 +33,29 @@ const useCategorySettings = ({ onDeleteCandidate }: CategorySettingsProps) => {
           description: "Deleted Category successfuly",
           placement: "bottomLeft",
         });
+      } else if (type === "sub-category" && deleteCandidate) {
+        deleteSubCategory(deleteCandidate._id);
+        setDeleteConfirmationOpen(false);
+        ref.current?.reload();
+        notification.success({
+          message: `Success`,
+          description: "Deleted sub-Category successfuly",
+          placement: "bottomLeft",
+        });
+      } else if (type === "main-category" && deleteCandidate) {
+        deleteMainCategory(deleteCandidate._id);
+        setDeleteConfirmationOpen(false);
+        ref.current?.reload();
+        notification.success({
+          message: `Success`,
+          description: "Deleted main-Category successfuly",
+          placement: "bottomLeft",
+        });
       }
     } catch (error) {
       Modal.warning({
         title: "Error",
-        content: "Failed to delete the category",
+        content: `Failed to delete the ${type}`,
       });
     }
   };
